# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/posts": {
      "get": {
        "operationId": "PostsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "course_tag",
            "required": false,
            "in": "query",
            "schema": {
              "example": "CS201",
              "type": "string"
            }
          },
          {
            "name": "feed",
            "required": false,
            "in": "query",
            "description": "following = เฉพาะกระทู้ของคนที่ฉันติดตาม (รวมของตัวเอง)",
            "schema": {
              "default": "all",
              "type": "string",
              "enum": [
                "all",
                "following"
              ]
            }
          },
          {
            "name": "author_username",
            "required": false,
            "in": "query",
            "description": "เฉพาะกระทู้ของคนนี้ — ใช้ทำหน้าโปรไฟล์",
            "schema": {
              "example": "6704101382-anuchat",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PostResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "กระดานข่าวทั้งหมด กรองตามแท็กวิชาได้",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ตั้งกระทู้ใหม่",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/v1/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบกระทู้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายละเอียดกระทู้",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "ลบกระทู้ของคนอื่นไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบกระทู้ (เจ้าของ บุคลากร หรือผู้ดูแล)",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/v1/posts/{id}/comments": {
      "get": {
        "operationId": "PostsController_listComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PostCommentResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ความคิดเห็นในกระทู้ เรียงเก่าไปใหม่",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostsController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostCommentResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ตอบกระทู้",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/v1/posts/{id}/comments/{commentId}": {
      "delete": {
        "operationId": "PostsController_removeComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "ลบความคิดเห็นของคนอื่นไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบความคิดเห็น",
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "CS Nexus — aie4-social-reels",
    "description": "ระบบย่อยภายใต้ CSMJU2030 · ตัวตนมาจาก API Gateway ผ่าน header X-User-Id / X-Layer1-Role / X-Faculty",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "number",
            "example": 2
          },
          "per_page": {
            "type": "number",
            "example": 50
          },
          "total_pages": {
            "type": "number",
            "example": 50
          },
          "total_items": {
            "type": "number",
            "example": 2480
          }
        },
        "required": [
          "current_page",
          "per_page",
          "total_pages",
          "total_items"
        ]
      },
...
      "EmojiCountDto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string",
            "example": "👍"
          },
          "count": {
            "type": "number",
            "example": 12
          },
          "reacted_by_me": {
            "type": "boolean",
            "description": "ฉันกดอิโมจินี้อยู่ไหม"
          }
        },
        "required": [
          "emoji",
          "count",
          "reacted_by_me"
        ]
      },
      "ReactionSummaryDto": {
        "type": "object",
        "properties": {
          "target_kind": {
            "type": "string",
            "enum": [
              "MESSAGE",
              "POST",
              "REEL"
            ]
          },
          "target_id": {
            "type": "string"
          },
          "totals": {
            "description": "เรียงจากยอดมากไปน้อย เอาเฉพาะอิโมจิที่มีคนกดจริง",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmojiCountDto"
            }
          },
          "total_count": {
            "type": "number",
            "example": 15,
            "description": "ผลรวมทุกอิโมจิ"
          }
        },
        "required": [
          "target_kind",
          "target_id",
          "totals",
          "total_count"
        ]
      },
      "PostResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "course_tag": {
            "type": "object",
            "nullable": true
          },
          "author_username": {
            "type": "string",
            "example": "6704101382-anuchat"
          },
          "comment_count": {
            "type": "number"
          },
          "reactions": {
            "nullable": true,
            "description": "ยอดอิโมจิรีแอ็กชัน พร้อมบอกว่าผู้เรียกกดอะไรไว้",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReactionSummaryDto"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "example": "2026-08-11T09:30:00+07:00"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "course_tag",
          "author_username",
          "comment_count",
          "reactions",
          "created_at"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "ถามเรื่อง pointer ใน C ครับ"
          },
          "content": {
            "type": "string",
            "example": "ทำไม *ptr กับ &var ให้ผลไม่เหมือนกันครับ"
          },
          "course_tag": {
            "type": "string",
            "example": "CS201",
            "description": "แท็กวิชา ใช้กรองกระดานข่าวตามรายวิชา"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "PostCommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "post_id": {
            "type": "string"
          },
          "author_username": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "post_id",
          "author_username",
          "content",
          "created_at"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "ลองวาดภาพหน่วยความจำดูครับ จะเห็นชัดขึ้น"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}